+++ client/src/components/Modal/ChangeModal.tsx
@@ -1,11 +1,7 @@
 import React, { useState } from 'react';
-import {
-  Button,
-  Dialog,
-  DialogBody,
-  DialogFooter,
-  Input,
-} from '@material-tailwind/react';
+import { ChangeEvent } from 'react';
+import { Button, Dialog, DialogBody, DialogFooter, Input } from '@material-tailwind/react';
+import axios from 'axios';
 
 type props = {
   title: string;
@@ -13,6 +9,9 @@ type props = {
   warning: string;
   btnName: string;
   btnCheck: string;
+  changeItem: string;
+  userId: number;
+  API: string;
 };
 
 export const ChangeModal = ({
@@ -21,20 +20,46 @@ export const ChangeModal = ({
   warning,
   btnName,
   btnCheck,
+  changeItem,
+  userId,
+  API,
 }: props) => {
   const [open, setOpen] = useState<boolean>(false);
+  const [inputValue, setInputValue] = useState<string>('');
 
-  const handleOpen = () => setOpen(!open);
+  console.log(userId, API);
+
+  const handleNameChange = async (newName: string) => {
+    try {
+      const data = {
+        [changeItem]: newName,
+      };
+      const response = await axios.patch(`http://localhost:8080/member/${userId}`, data);
+      console.log(response.data);
+    } catch (error) {
+      console.log(`$changeItem}`, error);
+    }
+  };
+
+  const handleOpen = () => {
+    setOpen(!open);
+  };
+
+  const handleClick = () => {
+    handleNameChange(inputValue);
+    handleOpen();
+  };
+
+  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
+    setInputValue(e.target.value);
+  };
 
   return (
     <>
-      <Button onClick={handleOpen}>{btnName}</Button>
-      <Dialog
-        open={open}
-        handler={handleOpen}
-        size="xs"
-        className="overflow-hidden "
-      >
+      <Button onClick={handleOpen} size="sm">
+        {btnName}
+      </Button>
+      <Dialog open={open} handler={handleOpen} size="xs" className="overflow-hidden">
         <DialogBody divider>
           <p className="text-center text-black">{title}</p>
           <div className="grid grid-flow-col">
@@ -42,12 +67,14 @@ export const ChangeModal = ({
               label={text}
               crossOrigin={undefined}
               color="blue"
-              className="text-black "
+              className="text-black"
+              value={inputValue}
+              onChange={handleChange} // 입력 값을 업데이트
             />
             <Button
               variant="outlined"
               color="red"
-              onClick={handleOpen}
+              onClick={handleClick}
               className="col-span-1 p-2 ml-5"
             >
               {btnCheck}
